Refactored home, login and registration pages to use modular components for improved readability and maintainability, removed redudant code.
