+++ app/login/page.tsx
@@ -1,142 +1,38 @@
-"use client"
+"use client";
 
-import { useState, useEffect } from "react"
-import Link from "next/link"
-import { useRouter } from "next/navigation"
-import { zodResolver } from "@hookform/resolvers/zod"
-import { useForm } from "react-hook-form"
-import * as z from "zod"
-import { Button } from "@/components/ui/button"
-import { Input } from "@/components/ui/input"
-import { Label } from "@/components/ui/label"
-import { Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle } from "@/components/ui/card"
-import { Alert, AlertDescription } from "@/components/ui/alert"
-import { AlertCircle, Zap } from "lucide-react"
-import { useAuth } from "@/contexts/auth-context"
+import { useEffect } from "react";
+import { useRouter } from "next/navigation";
+import { useAuth } from "@/contexts/auth-context";
+import { Card } from "@/components/ui/card";
+import { LogoHeader } from "@/components/logo-header";
+import { LoginForm } from "@/components/login/login-form";
 
-const loginSchema = z.object({
-  username: z.string().min(3, { message: "Please enter a valid username" }),
-  password: z.string().min(6, { message: "Password must be at least 6 characters" }),
-})
-
-type LoginFormValues = z.infer<typeof loginSchema>
-
+// Login page with authentication redirect and modular components
 export default function LoginPage() {
-  const router = useRouter()
-  const { login, isAuthenticated, isAdmin} = useAuth()
-  const [error, setError] = useState<string | null>(null)
-  const [isLoading, setIsLoading] = useState<boolean>(false)
-  const {
-    register,
-    handleSubmit,
-    formState: { errors },
-  } = useForm<LoginFormValues>({
-    resolver: zodResolver(loginSchema),
-    defaultValues: {
-      username: "",
-      password: "",
-    },
-  })
+  const router = useRouter();
+  const { isAuthenticated, isAdmin } = useAuth();
 
   useEffect(() => {
-    // Simpler redirect logic - just check if user is authenticated
+    // Redirect if already authenticated
     if (isAuthenticated) {
       if (isAdmin) {
-        router.push('/admin/dashboard');
+        router.push("/admin/dashboard");
       } else {
-        router.push('/dashboard');
+        router.push("/dashboard");
       }
     }
   }, [isAuthenticated, isAdmin, router]);
 
-  const onSubmit = async (data: LoginFormValues) => {
-    setIsLoading(true)
-    setError(null)
-
-    try {
-      const response = await login(data)
-      
-      // Redirect based on user role
-      if (response.user.role === 'admin') {
-        router.push('/admin/dashboard')
-      } else {
-        router.push('/dashboard')
-      }
-    } catch (err: any) {
-      
-      setError(err.response?.data?.message || "Failed to login. Please check your credentials.")
-    } finally {
-      setIsLoading(false)
-    }
-  }
-
   return (
     <div className="flex min-h-screen flex-col items-center justify-center bg-black px-4 py-12">
+      {/* Logo/Header */}
       <div className="absolute top-4 left-4">
-        <Link href="/" className="flex flex-col items-start text-white">
-          <div className="flex items-center">
-            <Zap className="h-6 w-6 text-yellow-400 mr-2" />
-            <span className="font-bold text-xl">Flash</span>
-          </div>
-          <span className="text-xs text-gray-400 ml-8">by QuantHive</span>
-        </Link>
+        <LogoHeader />
       </div>
+      {/* Login Card */}
       <Card className="w-full max-w-md bg-gray-900 text-white border-gray-800">
-        <CardHeader className="space-y-1">
-          <CardTitle className="text-2xl font-bold">Sign in to your account</CardTitle>
-          <CardDescription className="text-gray-400">
-            Enter your username and password to sign in
-          </CardDescription>
-        </CardHeader>
-        <CardContent>
-          {error && (
-            <Alert variant="destructive" className="mb-4">
-              <AlertCircle className="h-4 w-4" />
-              <AlertDescription>{error}</AlertDescription>
-            </Alert>
-          )}
-
-          <form onSubmit={handleSubmit(onSubmit)} className="space-y-4">
-            <div className="space-y-2">
-              <Label htmlFor="username" className="text-white">
-                Username
-              </Label>
-              <Input
-                id="username"
-                placeholder="johndoe123"
-                {...register("username")}
-                autoComplete="username"
-                className="bg-gray-800 border-gray-700 text-white"
-              />
-              {errors.username && <p className="text-sm text-red-500">{errors.username.message}</p>}
-            </div>
-            <div className="space-y-2">
-              <Label htmlFor="password" className="text-white">
-                Password
-              </Label>
-              <Input
-                id="password"
-                type="password"
-                {...register("password")}
-                autoComplete="current-password"
-                className="bg-gray-800 border-gray-700 text-white"
-              />
-              {errors.password && <p className="text-sm text-red-500">{errors.password.message}</p>}
-            </div>
-            <Button type="submit" className="w-full bg-yellow-400 text-black hover:bg-yellow-500" disabled={isLoading}>
-              {isLoading ? "Signing in..." : "Sign In"}
-            </Button>
-          </form>
-        </CardContent>
-        <CardFooter className="flex flex-col">
-          <div className="text-sm text-gray-400 mt-2">
-            Don't have an account?{" "}
-            <Link href="/register" className="text-yellow-400 hover:underline">
-              Register
-            </Link>
-          </div>
-        </CardFooter>
+        <LoginForm />
       </Card>
     </div>
-  )
-} 
+  );
+}
