+++ components/login/login-form.tsx
@@ -1,0 +1,140 @@
+"use client";
+
+import { useState } from "react";
+import { useRouter } from "next/navigation";
+import { zodResolver } from "@hookform/resolvers/zod";
+import { useForm } from "react-hook-form";
+import * as z from "zod";
+import { Button } from "@/components/ui/button";
+import { Input } from "@/components/ui/input";
+import { Label } from "@/components/ui/label";
+import {
+  CardContent,
+  CardFooter,
+  CardHeader,
+  CardTitle,
+  CardDescription,
+} from "@/components/ui/card";
+import { Alert, AlertDescription } from "@/components/ui/alert";
+import { AlertCircle } from "lucide-react";
+import Link from "next/link";
+import { useAuth } from "@/contexts/auth-context";
+
+const loginSchema = z.object({
+  username: z.string().min(3, { message: "Please enter a valid username" }),
+  password: z
+    .string()
+    .min(6, { message: "Password must be at least 6 characters" }),
+});
+
+type LoginFormValues = z.infer<typeof loginSchema>;
+
+/**
+ * LoginForm component for the login page.
+ */
+export function LoginForm() {
+  const router = useRouter();
+  const { login } = useAuth();
+  const [error, setError] = useState<string | null>(null);
+  const [isLoading, setIsLoading] = useState<boolean>(false);
+  const {
+    register,
+    handleSubmit,
+    formState: { errors },
+  } = useForm<LoginFormValues>({
+    resolver: zodResolver(loginSchema),
+    defaultValues: {
+      username: "",
+      password: "",
+    },
+  });
+
+  const onSubmit = async (data: LoginFormValues) => {
+    setIsLoading(true);
+    setError(null);
+    try {
+      const response = await login(data);
+      // Redirect based on user role
+      if (response.user.role === "admin") {
+        router.push("/admin/dashboard");
+      } else {
+        router.push("/dashboard");
+      }
+    } catch (err: any) {
+      setError(
+        err.response?.data?.message ||
+          "Failed to login. Please check your credentials."
+      );
+    } finally {
+      setIsLoading(false);
+    }
+  };
+
+  return (
+    <>
+      <CardHeader className="space-y-1">
+        <CardTitle className="text-2xl font-bold">
+          Sign in to your account
+        </CardTitle>
+        <CardDescription className="text-gray-400">
+          Enter your username and password to sign in
+        </CardDescription>
+      </CardHeader>
+      <CardContent>
+        {error && (
+          <Alert variant="destructive" className="mb-4">
+            <AlertCircle className="h-4 w-4" />
+            <AlertDescription>{error}</AlertDescription>
+          </Alert>
+        )}
+        <form onSubmit={handleSubmit(onSubmit)} className="space-y-4">
+          <div className="space-y-2">
+            <Label htmlFor="username" className="text-white">
+              Username
+            </Label>
+            <Input
+              id="username"
+              placeholder="johndoe123"
+              {...register("username")}
+              autoComplete="username"
+              className="bg-gray-800 border-gray-700 text-white"
+            />
+            {errors.username && (
+              <p className="text-sm text-red-500">{errors.username.message}</p>
+            )}
+          </div>
+          <div className="space-y-2">
+            <Label htmlFor="password" className="text-white">
+              Password
+            </Label>
+            <Input
+              id="password"
+              type="password"
+              {...register("password")}
+              autoComplete="current-password"
+              className="bg-gray-800 border-gray-700 text-white"
+            />
+            {errors.password && (
+              <p className="text-sm text-red-500">{errors.password.message}</p>
+            )}
+          </div>
+          <Button
+            type="submit"
+            className="w-full bg-yellow-400 text-black hover:bg-yellow-500"
+            disabled={isLoading}
+          >
+            {isLoading ? "Signing in..." : "Sign In"}
+          </Button>
+        </form>
+      </CardContent>
+      <CardFooter className="flex flex-col">
+        <div className="text-sm text-gray-400 mt-2">
+          Don't have an account?{" "}
+          <Link href="/register" className="text-yellow-400 hover:underline">
+            Register
+          </Link>
+        </div>
+      </CardFooter>
+    </>
+  );
+}
